requirements.txt
dotenv
flask
PyPDF2
groq
python-docx


